`yamale.schema.util` moved to `yamale.util`.
The general-purpose `yamale.schema.util` module has been moved to the root
`yamale` package for use elsewhere (e.g., the `yamale.validators` subpackage).
