+++ yamale/util.py
